requirements.txt
Django==1.11.4
djangorestframework==3.6.3
Pillow==4.2.1
django-filter==1.1.0
